Make footer component reusable.
Moved footer component into common folder component then modified
the style to make it reusable across different pages.
