+++ src/pages/NotFound.jsx
@@ -1,12 +1,13 @@
-import { Nav } from '../components/common';
+import { Nav, Footer } from '../components/common';
 
 const NotFound = () => {
   return (
     <main className="relative">
       <Nav />
-      <section className="padding max-container">
+      <section className="padding max-container min-h-screen">
         <h1>NotFound</h1>
       </section>
+      <Footer />
     </main>
   );
 };
